+++ ros1_ign_gazebo_demos/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package ros1_ign_gazebo_demos
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.4.0 (2019-07-16)
+------------------
+
 0.3.1 (2019-07-01)
 ------------------
 * Merge pull request `#24 <https://github.com/osrf/ros1_ign_bridge/issues/24>`_ from osrf/fix_dep